Push whole impls down into common using typealias trick
