+++ samples/star/src/commonJvm/kotlin/com/slack/circuit/star/di/InjectAnnotations.jvm.kt
@@ -1,0 +1,16 @@
+// Copyright (C) 2024 Slack Technologies, LLC
+// SPDX-License-Identifier: Apache-2.0
+package com.slack.circuit.star.di
+
+import dagger.assisted.Assisted
+import dagger.assisted.AssistedFactory
+import dagger.assisted.AssistedInject
+import javax.inject.Inject
+
+actual typealias Inject = Inject
+
+actual typealias AssistedInject = AssistedInject
+
+actual typealias Assisted = Assisted
+
+actual typealias AssistedFactory = AssistedFactory
